pyproject.toml
[project]
name = "stock-ana"
version = "0.1.0"
description = "股票技术分析与筛选工具"
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    # 数据获取
    "akshare>=1.14.0",        # A股数据（也用于美股/港股新浪源）
    "futu-api>=9.0.0",        # 富途 OpenD（港股/美股/A股）

    # 数据处理与分析
    "pandas>=2.1.0,<3.0.0",
    "numpy>=1.26.0,<2.5.0",
    "ta>=0.11.0",             # 技术指标库（纯 Python，跨平台无障碍）

    # 可视化
    "mplfinance>=0.12.9b7",   # K线图（最新稳定版为 pre-release）
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",         # 交互式图表

    # 工具
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",   # 环境变量管理
    "loguru>=0.7.0",          # 日志

    # AI 分析
    "gemini-webapi>=1.18.0",  # Gemini Web 逆向 API
    "browser-cookie3>=0.20.0", # 自动读取浏览器 Cookie
    "python-docx>=1.0.0",     # Word 文档输出
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.5.0",            # 代码检查与格式化
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
